requirements.txt
Django>=4.2
fastapi
uvicorn
python-socketio
simple-websocket